MutableIntHistogram
